Se crearon las capas de Modelo y controlador para los CRUD

+++ TeamDigitalYelmos/src/main/java/com/Yelmos/TeamDigitalYelmos/Yelmos2/Categoria.java
@@ -16,7 +16,7 @@ public class Categoria implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "idCategoria")
-	private int idCatego;
+	private Long idCatego;
 	
 	@Column(name = "nombreCategoria", length = 40)
 	private String nomCtego;
